Readme improvements
ncurses and libc are installed by default on OS X.
The runtime automatically looks into /usr/lib and /usr/local/lib, so there's no need to symlink libc and libncurses. Specify that libtermkey needs to be copied in the output directory if the target machine isn't guaranteed to have it installed.